requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.0
google-genai>=1.0.0
rich>=13.7.0
streamlit>=1.38.0